refactor routes for enterprise and points
